Refactor monolithic Assessment, Copilot, and Simulator components
