+++ client/src/components/assessment/types.ts
@@ -1,0 +1,10 @@
+export interface RiskFactor {
+  name: string;
+  impact: "positive" | "negative" | string;
+  description: string;
+}
+
+export interface FactorBreakdown extends RiskFactor {
+  strength: number;
+  plainReason: string;
+}
